refactor : 문제 번호 Long -> Integer로 변경
문제 개수가 Integer 한계 (2^31)까지 가지 않을 것이기에 Integer로 변경

+++ DCBA/src/main/java/kingrangE/DCBA/domain/Exercise.java
@@ -1,17 +1,17 @@
 package kingrangE.DCBA.domain;
 
 public class Exercise {
-    private Long num;
+    private Integer num;
     private String title;
     private String best_answer;
     private Subject subject;
     private Level level;
 
-    public Long getNum() {
+    public Integer getNum() {
         return num;
     }
 
-    public void setNum(Long num) {
+    public void setNum(Integer num) {
         this.num = num;
     }
